refactor: update file pathes
